requirements.txt
openai
python-dotenv
sounddevice
numpy
requests
